web_apps/logsworth/requirements.txt
Flask==0.10.1
Flask-HTTPAuth==3.1.2
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
SQLAlchemy==1.0.12
Werkzeug==0.11.9
wsgiref==0.1.2
